Final creation User
When a user is logging, it is creating an access token throught /login.

No need to login to add User and then type the link /oauth/v2/token

The whole process is manged in UserController.

Adding a param in Client "name" to make it more commercial and do a demo nicely without showing the whole token of the customer. Can easily be changed.

It will be updated to be sure that both add user work through ClientName or Client ID / Secret